% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/da2fya.R
\name{da2fya}
\alias{da2fya}
\title{produce a matrix containing the age-specific distribution of remaining lifespans}
\usage{
da2fya(da, stagger = FALSE)
}
\arguments{
\item{da}{a vector of single age lifetable deaths (dx).}

\item{stagger}{logical. Should the da vector be element-wise averaged to account for deaths being spread through the year?}
}
\value{
a matrix of the conditional distribution of remaining lifetime: chronological age in rows and remaining years in columns.
}
\description{
Given a single-age vector of lifetable deaths (any radix), we convert to a matrix with chronological age in rows and thanatological age in columns. Each row represents the distribution of remaining lifespans for the given chronological age, and is scaled to sum to 1.
}
\details{
if \code{stagger} is \code{TRUE}, age 0 is not treated specially at this time, even though most of \eqn{d_0} is concentrated in the first day and weeks.
}
