% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{dx2lx}
\alias{dx2lx}
\title{dx2lx a simple conversion function}
\usage{
dx2lx(dx, radix = 1)
}
\arguments{
\item{dx}{the lifetable dx column, any radix ok}

\item{radix,}{the destination radix, l(0)}
}
\value{
lx a survival function, with l(0) set to radix
}
\description{
Functionalized for the sake of modularity. No NA checking done.
}
\details{

}

