% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Lmoments.R
\name{getB0b_ta}
\alias{getB0b_ta}
\title{Adam's function getB0b, reworked as a function of age}
\usage{
getB0b_ta(dx, age)
}
\arguments{
\item{dx}{vector the lifetable deaths distribution}

\item{age}{vector of (single) completed ages}
}
\value{
age-specific vector of ...
}
\description{
Awaiting description from Adam. This amounts to a dx-weighted average remaining years of life, approximately equal to ex. No radix argument required here.
}

