% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/momentN.R
\name{getex}
\alias{getex}
\title{calculate remaining life expectancy as the dx-weighted mean of remaining lifetimes}
\usage{
getex(dx, ax = rep(0.5, length(dx)))
}
\arguments{
\item{dx}{a vector of single age lifetable deaths (dx). Any radix.}

\item{ax}{Chiang's lifetable ax.}
}
\value{
vector of alternative calculation of ex
}
\description{
This should provide an estimate rather close to the typical lifetable ex, calculated as Tx/lx. Here we weight the lengths of remaining lifelines by the lifetable deaths distribution (i.e. integration in the opposite direction).
}

