% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Lmoments.R
\name{getL2b_ta}
\alias{getL2b_ta}
\title{Adam's function getL2b, reworked as a function of age}
\usage{
getL2b_ta(dx, age, radix = 1e+05)
}
\arguments{
\item{dx}{vector the lifetable deaths distribution}

\item{age}{vector of (single) completed ages}

\item{radix}{for scaling. Should be reasonably large, not 1 ...}
}
\value{
vector of age specific 2nd L-moments
}
\description{
Awaiting further description from Adam. calculates age-specific 2nd L-moment.
}

