% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Lmoments.R
\name{getL3b_ta}
\alias{getL3b_ta}
\title{Adam's function getL3, reworked as a function of age}
\usage{
getL3b_ta(dx, age, radix = 1e+05)
}
\arguments{
\item{dx}{vector the lifetable deaths distribution}

\item{age}{vector of (single) completed ages}

\item{radix}{for scaling. Should be reasonably large, not 1 ...}
}
\value{
vector of age specific 3rd L-moments
}
\description{
Calculates age-specific 3rd L-Moment. Awaiting further description from Adam.
}

