% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Lmoments.R
\name{getLCV_ta}
\alias{getLCV_ta}
\title{Adam's function getLCV, reworked as a function of age}
\usage{
getLCV_ta(dx, age, radix = 1e+05)
}
\arguments{
\item{dx}{vector the lifetable deaths distribution}

\item{age}{vector of (single) completed ages}

\item{radix}{for scaling. Should be reasonably large, not 1 ...}
}
\value{
vector of age specific L-CV
}
\description{
Calculates coefficient of L-variation, identical to Gini coefficient. Ratio of second to first L-moments.
}

