% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Lmoments.R
\name{getLSkew_ta}
\alias{getLSkew_ta}
\title{Adam's function getLSkew, reworked as a function of age}
\usage{
getLSkew_ta(dx, age, radix = 1e+05)
}
\arguments{
\item{dx}{vector the lifetable deaths distribution}

\item{age}{vector of (single) completed ages}

\item{radix}{for scaling. Should be reasonably large, not 1 ...}
}
\value{
vector of age specific L-skewness
}
\description{
Ratio of third to second L-moments gives the skewness. This function calculates the L-skewness of the distribution of remaining years of life as a function of age.
}

