% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getMode}
\alias{getMode}
\title{calculate the conditional modal remaining years of life}
\usage{
getMode(dx)
}
\arguments{
\item{dx}{a vector of single age lifetable deaths, any radix}
}
\value{
numeric vector of modal remaining lifespans.
}
\description{
This function uses a spline function fit to the conditional remaining lifetime function to approximate the modal remaining years of life for each age. This function might not be robust. Note, results can seem like a function of discrete segments because a mode can be the same for many ages, getting 1 year closer each age. It's less aesthetically interesting than the median remaining lifespan, but still potentially interesting. After the primary old age mode, results get choppy due to noise.
}
\details{

}

