% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedianMode.R
\name{getQuantile}
\alias{getQuantile}
\title{calculate quantiles of the conditional distribution of remaining years of life}
\usage{
getQuantile(dx, probs = c(0.1, 0.25, 0.5, 0.75, 0.9))
}
\arguments{
\item{dx}{a vector of single age lifetable deaths, any radix}

\item{probs}{a vector of quantiles}
}
\value{
a vector or matrix of quantiles from the distribution of remaining lifespans.
}
\description{
This function uses a monotonic spline function fit to the cdf to approximate quantiles from the remaining years of life for each age. Final values may be iffy, you may wish to truncate after, say, omega - 5.
}
