% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/momentN.R
\name{getSkewst}
\alias{getSkewst}
\title{calculate the conditional skewness of remaining years of life.}
\usage{
getSkewst(dx, ax)
}
\arguments{
\item{dx}{a vector of single age lifetable deaths (dx). Any radix.}

\item{ax}{Chiang's lifetable ax}
}
\value{
vector of the age-specific skewness of the conditional distribution of remaining years of life
}
\description{
Ratio of the 3rd to 2nd moments.
}

