% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/da2fya.R
\name{MinfNA}
\alias{MinfNA}
\title{replace Inf elements with NA}
\usage{
MinfNA(M)
}
\arguments{
\item{M}{a vector or matrix}
}
\value{
M, with Inf imputed with NAs.
}
\description{
replace Inf elements with NA
}
