% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/da2fya.R
\name{Mna0}
\alias{Mna0}
\title{replace NAs and NaNs with 0}
\usage{
Mna0(M)
}
\arguments{
\item{M}{a vector or matrix}
}
\value{
M, with NAs and NaNs imputed with zeros.
}
\description{
replace NAs and NaNs with 0
}
