% Generated by roxygen2 (4.0.1): do not edit by hand
\name{momentN}
\alias{momentN}
\title{calculate the Nth moment of remaining years of life as a function of age}
\usage{
momentN(dx, n, ax = rep(0.5, length(dx)))
}
\arguments{
\item{dx}{a vector of single age lifetable deaths (dx). Any radix.}

\item{n}{which moment?}

\item{ax}{Chiang's lifetable ax.}
}
\value{
vector of age specific nth moments.
}
\description{
The Nth moment of the conditional distribution of remaining years of life, \eqn{\eta _n(y|a)} is defined as:
\deqn{\eta _n(y|a) =  \int_{y=0}^\infty (y-e(a))^n f(y|a) \text{d}y}
ax is used as an argument to account for infants. If missing, assumed as .5. Ages assumed single.
}

