% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lmoments.R
\name{getB1b_ta}
\alias{getB1b_ta}
\title{Adam's function getB1b, reworked as a function of age}
\usage{
getB1b_ta(dx, age, radix = 1e+05)
}
\arguments{
\item{dx}{vector the lifetable deaths distribution}

\item{age}{vector of (single) completed ages}

\item{radix}{for scaling. Should be reasonably large, not 1 ...}
}
\value{
vector of age-specific ...
}
\description{
Awaiting description from Adam.
}
