% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentN.R
\name{getKurtst}
\alias{getKurtst}
\title{calculate the conditional kurtosis of remaining years of life.}
\usage{
getKurtst(dx, ax)
}
\arguments{
\item{dx}{a vector of single age lifetable deaths (dx). Any radix.}

\item{ax}{Chiang's lifetable ax.}
}
\value{
vector of the age-specific kurtosis of the conditional distribution of remaining years of life
}
\description{
Ratio of the 4th to 2nd moments.
}
