% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedianMode.R
\name{getMedian}
\alias{getMedian}
\title{calculate the conditional median remaining years of life}
\usage{
getMedian(dx)
}
\arguments{
\item{dx}{a vector of single age lifetable deaths, any radix}
}
\value{
numeric vector of median remaining lifespans.
}
\description{
This function uses a monotonic spline function fit to the cdf to approximate the median remaining years of life for each age. Final values may be iffy, you may wish to truncate after, say, omega - 5.
}
